+++ tests/testthat/helper-dplyr.R
@@ -1,0 +1,1 @@
+library(dplyr, warn.conflicts = FALSE)